fixed particle cluster and put compute_forces in interaction.py
